+++ guardian-ui/src/components/SetConfiguration.tsx
@@ -14,8 +14,8 @@ import { GuardianRole, Network } from '../types';
 import { ReactComponent as FedimintLogo } from '../assets/svgs/fedimint.svg';
 import { ReactComponent as BitcoinLogo } from '../assets/svgs/bitcoin.svg';
 import { ReactComponent as ArrowRightIcon } from '../assets/svgs/arrow-right.svg';
-import { FormGroup } from './FormGroup';
-import { FormGroupHeading } from './FormGroupHeading';
+import { FormGroup } from './ui/FormGroup';
+import { FormGroupHeading } from './ui/FormGroupHeading';
 
 interface Props {
   next(): void;
